#35 Dissolve _utils.py grab-bag into cohesive modules
_utils.py mixed four unrelated concerns. Redistribute its functions to
homes that match what they do, and delete the module:

  search_ast_type                 -> _expressions.py (AST helpers)
  to_cpp_name/to_camel_case/
    generate_header_guard         -> _names.py (C++ identifier naming)
  get_compartment_size/
    get_function_definition_args   -> _sbml_reader.py (renamed from
                                     _sbml_loader.py: SBML loading + readers)

Naming tests move alongside (test_utils.py -> test_names.py). Also drop the
resolved "Add priority" TODO in _format_events and correct the trigger
distance multi-operand comment.

Pure move: reference outputs regenerate unchanged.
