+++ chaste_sbml/tests/test_names.py
@@ -8,8 +8,11 @@ from chaste_sbml._names import (
     CPP_KEYWORDS,
     NameManager,
     find_name_conflicts,
+    generate_header_guard,
     is_valid_cpp_identifier,
     resolve_cpp_name,
+    to_camel_case,
+    to_cpp_name,
     unique_name,
 )
 
@@ -161,3 +164,63 @@ def test_name_manager_resolve_leaves_safe_ids_untouched():
     NameManager(sbml_model).resolve_real_id_conflicts()
 
     assert sbml_model.getElementBySId("cell") is not None
+
+
+def test_generate_header_guard():
+    """Test header guard generation."""
+    test_cases = [
+        ("", ""),
+        (" ", ""),
+        ("foo.hpp", "FOO_HPP_"),
+        ("foo_bar.hpp", "FOO_BAR_HPP_"),
+        ("fooBar.hpp", "FOO_BAR_HPP_"),
+        ("FooBar.hpp", "FOO_BAR_HPP_"),
+        ("Foo100bar.hpp", "FOO_100BAR_HPP_"),
+        ("Foo200Bar.hpp", "FOO_200_BAR_HPP_"),
+        ("FooB300ar.hpp", "FOO_B300AR_HPP_"),
+        ("400FooBar.hpp", "_400_FOO_BAR_HPP_"),
+        ("TysonNovak2001.hpp", "TYSON_NOVAK_2001_HPP_"),
+        ("TestSemantic00001L2V5Sbml.hpp", "TEST_SEMANTIC_00001_L2_V5_SBML_HPP_"),
+    ]
+
+    for test_input, expected_output in test_cases:
+        assert generate_header_guard(test_input) == expected_output
+
+
+def test_to_camel_case():
+    """Test variable name case staggering."""
+    test_cases = [
+        ("", ""),
+        (" ", ""),
+        ("foo", "foo"),
+        ("foo_bar", "fooBar"),
+        ("foo bar", "fooBar"),
+        ("foo.bar", "fooBar"),
+        ("abstract_foo_bar", "abstractFooBar"),
+        ("foo1bar", "foo1Bar"),
+        ("foo100bar", "foo100Bar"),
+        ("TysonNovak2001", "TysonNovak2001"),
+    ]
+
+    for test_input, expected_output in test_cases:
+        assert to_camel_case(test_input) == expected_output
+
+
+def test_to_cpp_name():
+    """Test variable name sanitization."""
+    test_cases = [
+        ("", ""),
+        (" ", ""),
+        ("foo", "foo"),
+        ("foo_bar", "foo_bar"),
+        ("foo___bar", "foo___bar"),
+        ("foo bar", "foo_bar"),
+        ("foo_ _bar", "foo___bar"),
+        ("foo_ . _bar", "foo_____bar"),
+        ("foo1bar", "foo1bar"),
+        ("1foo", "_1foo"),
+        ("foo100bar", "foo100bar"),
+    ]
+
+    for test_input, expected_output in test_cases:
+        assert to_cpp_name(test_input) == expected_output
